# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaja (Kantin Jawara) API
  description: |
    # Kaja API Documentation

    API for **Kaja (Kantin Jawara)** canteen management system - a comprehensive solution for restaurant and food court operations.

    ## 🚀 Features

    - **Authentication System** - Secure user registration, login, and profile management
    - **Multi-Role Support** - Admin, Seller (Penjual), and Buyer (Pembeli) roles
    - **Menu Management** - Complete CRUD operations for food items and categories
    - **Smart Cart System** - Automatic price calculations, merchant-specific carts, and real-time updates
    - **Transaction Processing** - Order creation, status tracking, and management
    - **Payment Integration** - Multiple payment methods (Cash, Bank Transfer, E-Wallet)
    - **Real-time Chat** - Transaction-based messaging between buyers and sellers
    - **File Upload Support** - Images and documents for chat and payment proof

    ## 🛒 Cart Management

    The cart system provides intelligent shopping cart functionality with automatic calculations:

    - **Merchant-specific carts** - Separate carts for each merchant/seller
    - **Automatic price calculation** - Total prices calculated automatically based on quantity and unit price
    - **Real-time updates** - Cart totals recalculated instantly when items are added, updated, or removed
    - **Smart item management** - Duplicate items are merged with updated quantities
...

    ## 📋 Changelog

...

...

...

    ```
...
    ```

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /change-password:
    put:
      tags:
        - 🔐 Authentication
      summary: Change password
      description: |
        Change the user's password. Requires current password for security.

        **Note:** After successful password change, all existing tokens will be invalidated.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
            examples:
              change_password:
                summary: Change password
                value:
                  current_password: oldpassword123
                  new_password: newpassword123
                  new_password_confirmation: newpassword123
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully. Please login again."
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /delete-account:
    delete:
      tags:
        - 🔐 Authentication
      summary: Delete user account
      description: |
        Permanently delete the user account. This action cannot be undone.

        **Security:** Requires password confirmation and explicit "DELETE" confirmation.
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
            examples:
              delete_account:
                summary: Delete account
                value:
                  password: password123
                  confirmation: DELETE
      responses:
        "200":
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account deleted successfully"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /transactions/{id}:
    get:
      tags:
        - 📦 Transaction Management
      summary: Get transaction details
      description: Get detailed information about a specific transaction
      operationId: getTransactionDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Transaction ID
          example: 1
      responses:
        "200":
          description: Transaction details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TransactionDetails"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - 🛒 Customer APIs
      summary: Cancel transaction
      description: Cancel a transaction (only if not yet paid or confirmed)
      operationId: cancelTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Transaction ID
          example: 1
      responses:
        "200":
          description: Transaction cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Transaction cancelled successfully"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoint.

        **Usage:** `Authorization: Bearer your_token_here`
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                message: "Unauthenticated"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              summary: Access denied
              value:
                message: "This action is unauthorized"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            not_found:
              summary: Resource not found
              value:
                message: "Resource not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validation_failed:
              summary: Validation failed
              value:
                message: "The given data was invalid."
                errors:
                  field_name: ["The field is required."]
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        role:
          type: string
          enum: [admin, penjual, pembeli]
          description: User role in the system
          example: "pembeli"
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          description: Email verification timestamp
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-07-23T10:00:00Z"

    # Request schemas
...
    ChangePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
        - new_password_confirmation
      properties:
        current_password:
          type: string
          description: Current password
          example: "oldpassword123"
        new_password:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)
          example: "newpassword123"
        new_password_confirmation:
          type: string
          description: New password confirmation
          example: "newpassword123"
    DeleteAccountRequest:
      type: object
      required:
        - password
        - confirmation
      properties:
        password:
          type: string
          description: Current password for verification
          example: "password123"
        confirmation:
          type: string
          enum: ["DELETE"]
          description: Must be exactly "DELETE" to confirm
          example: "DELETE"

    # Category schemas
    Category:
      type: object
      properties:
        id:
          type: integer
          description: Category ID
          example: 1
        name:
          type: string
          description: Category name
          example: "Makanan Berat"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
...
    Menu:
      type: object
      properties:
        id:
          type: integer
          description: Menu ID
          example: 1
        name:
          type: string
          description: Menu item name
          example: "Nasi Goreng Spesial"
        description:
          type: string
          nullable: true
          description: Menu item description
          example: "Nasi goreng dengan telur dan ayam"
        price:
          type: number
          format: float
          description: Price in IDR
          example: 25000
        stock:
          type: integer
          description: Available stock
          example: 10
        image_url:
          type: string
          nullable: true
          description: Menu item image URL
          example: "https://example.com/nasigoreng.jpg"
        category:
          $ref: "#/components/schemas/Category"
        merchant:
          type: object
          properties:
            id:
              type: integer
              example: 2
            name:
              type: string
              example: "Warung Pak Joni"
        is_available:
          type: boolean
          description: Whether item is available (stock > 0)
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 2
        merchant_id:
          type: integer
          example: 1
        total_price:
          type: number
          format: float
          example: 50000
        status:
          type: string
          enum:
            [pending, paid, confirmed, preparing, ready, completed, cancelled]
          example: "pending"
        notes:
          type: string
          nullable: true
          example: "Pedas sedang, tanpa bawang"
        customer_name:
          type: string
          nullable: true
          example: "Budi Santoso"
        customer_phone:
          type: string
          nullable: true
          example: "081234567890"
        order_type:
          type: string
          enum: [takeaway, dine_in, delivery]
          nullable: true
          example: "takeaway"
        created_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
    TransactionDetails:
      allOf:
        - $ref: "#/components/schemas/Transaction"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/User"
            merchant:
              $ref: "#/components/schemas/User"
            items:
              type: array
              items:
                $ref: "#/components/schemas/TransactionItem"
            payment:
              $ref: "#/components/schemas/Payment"
              nullable: true
    TransactionItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        transaction_id:
          type: integer
          example: 1
        menu_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 25000
        total:
          type: number
          format: float
          example: 50000
        menu:
          $ref: "#/components/schemas/Menu"
        created_at:
          type: string
          format: date-time
          example: "2025-07-23T10:00:00Z"
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        transaction_id:
          type: integer
          example: 1
        amount:
          type: number
          format: float
          example: 50000
        method:
          type: string
          example: "Bank Transfer"
        status:
          type: string
          enum: [pending, verified, rejected]
          example: "pending"
        proof_url:
          type: string
          nullable: true
          example: "/storage/payment-proofs/proof_123.jpg"
        notes:
          type: string
          nullable: true
          example: "Transfer dari rekening BCA"
        verified_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-23T10:30:00Z"
        verified_by:
          type: integer
          nullable: true
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2025-07-23T10:15:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-07-23T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Something went wrong"
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Main error message
          example: "The given data was invalid."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
          example:
            email: ["The email field is required."]
            password: ["The password must be at least 8 characters."]
...